+++ ale/src/main/java/org/bytedeco/javacpp/presets/ale.java
@@ -39,10 +39,12 @@ import org.bytedeco.javacpp.tools.InfoMapper;
  */
 @Properties(
     value = {
-        @Platform(value = {"linux-x86", "macosx", "windows"}, compiler = "cpp11", link = "ale",
-            include = {"emucore/m6502/src/bspf/src/bspf.hxx", "emucore/Event.hxx", "common/Constants.h",
-                       "common/ColourPalette.hpp", "common/ScreenExporter.hpp", "environment/ale_ram.hpp",
-                       "environment/ale_screen.hpp", "environment/ale_state.hpp", "ale_interface.hpp"}),
+        @Platform(value = {"linux-x86", "macosx", "windows"}, compiler = "cpp11", define = "UNIQUE_PTR_NAMESPACE std", link = "ale",
+            include = {"emucore/m6502/src/bspf/src/bspf.hxx", "emucore/m6502/src/Device.hxx", "emucore/Control.hxx", "emucore/Event.hxx",
+                       "emucore/Random.hxx", "common/Constants.h", "common/Array.hxx", "common/display_screen.h", "emucore/M6532.hxx",
+                       "emucore/Cart.hxx", "emucore/Console.hxx", "emucore/Sound.hxx", "emucore/Settings.hxx", "emucore/OSystem.hxx",
+                       "common/ColourPalette.hpp", "common/ScreenExporter.hpp", "environment/ale_ram.hpp", "environment/ale_screen.hpp",
+                       "environment/ale_state.hpp", "environment/stella_environment_wrapper.hpp", "environment/stella_environment.hpp", "ale_interface.hpp"}),
         @Platform(value = "linux-x86",     preload = "SDL-1.2@.0", preloadpath = {"/usr/lib32/", "/usr/lib/"}),
         @Platform(value = "linux-x86_64",  preload = "SDL-1.2@.0", preloadpath = {"/usr/lib64/", "/usr/lib/"}),
         @Platform(value = "macosx-x86_64", preload = "SDL-1.2@.0", preloadpath = "/usr/local/lib/"),
@@ -51,8 +53,15 @@ import org.bytedeco.javacpp.tools.InfoMapper;
     target = "org.bytedeco.javacpp.ale")
 public class ale implements InfoMapper {
     public void map(InfoMap infoMap) {
-        infoMap.put(new Info("BSPF_strcasecmp", "BSPF_strncasecmp", "BSPF_snprintf", "BSPF_vsnprintf").cppTypes())
-               .put(new Info("ALEState::reset", "ALEInterface::theOSystem", "ALEInterface::theSettings", "ALEInterface::romSettings",
-                             "ALEInterface::environment", "ALEInterface::createOSystem", "ALEInterface::loadSettings").skip());
+        infoMap.put(new Info("DEBUGGER_SUPPORT", "CHEATCODE_SUPPORT").define(false))
+               .put(new Info("BSPF_strcasecmp", "BSPF_strncasecmp", "BSPF_snprintf", "BSPF_vsnprintf").cppTypes())
+               .put(new Info("Common::Array<Resolution>").pointerTypes("ResolutionList").define())
+               .put(new Info("StellaEnvironmentWrapper::m_environment").javaText("public native @MemberGetter @ByRef StellaEnvironment m_environment();"))
+               .put(new Info("StellaEnvironment::getWrapper").javaText("public native @Name(\"getWrapper().get\") StellaEnvironmentWrapper getWrapper();"))
+               .put(new Info("ALEInterface::theOSystem").javaText("public native @Name(\"theOSystem.get\") OSystem theOSystem();"))
+               .put(new Info("ALEInterface::theSettings").javaText("public native @Name(\"theSettings.get\") Settings theSettings();"))
+               .put(new Info("ALEInterface::romSettings").javaText("public native @Name(\"romSettings.get\") RomSettings romSettings();"))
+               .put(new Info("ALEInterface::environment").javaText("public native @Name(\"environment.get\") StellaEnvironment environment();"))
+               .put(new Info("AtariVox", "Common::Array<Resolution>::contains", "ALEState::reset").skip());
     }
 }